requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
idna==2.5
PyYAML==3.12
requests==2.18.1
six==1.10.0
slackclient==1.0.6
urllib3==1.21.1
websocket-client==0.42.1
